src/requirements.txt
disnake==2.2.2
python-dotenv==0.19.2